Split main module into client, config, cli
